FSA-5759: Created separated directories in helpers
