+++ projects/fsastorefrontlib/src/cms-components/checkout/components/add-options/add-options.module.ts
@@ -8,7 +8,7 @@ import { MediaModule, SpinnerModule } from '@spartacus/storefront';
 import { CartConnector } from '../../../../core/cart/connectors';
 import { FSCartService } from '../../../../core/cart/facade';
 import { AddOptionsComponent } from './add-options.component';
-import { SortModule } from '../../../../shared/util/helpers/sort.module';
+import { SortModule } from '../../../../shared/util/helpers/pipe/sort.module';
 
 @NgModule({
   imports: [
