+++ projects/fsastorefrontlib/src/shared/index.ts
@@ -1,5 +1,5 @@
 export * from './accordion/accordion-item.component';
 export * from './accordion/accordion.module';
-export * from './util/helpers/parseDate.pipe';
-export * from './util/helpers/formatDate.pipe';
-export * from './util/helpers/sortByName.pipe';
+export * from './util/helpers/pipe/parseDate.pipe';
+export * from './util/helpers/pipe/formatDate.pipe';
+export * from './util/helpers/pipe/sortByName.pipe';
